Made changes to allow execute of custom script from TUI
Basically I want to be able to do arbitrarily complex tasks
with the multitone system for testing in the lab but
it would be silly to add a bunch of custom definitions
to kidPy.py that are only for a one specific pupose
I added an example script of what can be done with this
It takes a fine target sweep and a larger target sweep
to measure cabel delay and gain varation.
In addition I added keywords to the target sweep that
lets me on the fly change the lo_step size and span.