added interactive plotting for targplot and tune_kids
I added an interactive plot option for targetplot so that
you have zoomed in plots of each resonator that you can
switch between with the right and left arrow keys.
I also added a similar functionality to the tune_kids.py
program. in that one you can also use the up down keys
to changed the number of look around points for finding
either the max of didq or the min of the resonator. This
is useful if you have some collisions and just want to
search only far enough. When I did this I also discovered
several bugs that effecivily made the tune_max_didq.py
not functional. In the past it was probably just still
tuning to the minimum of the resonator.