mostly added in clearing buffer of old packets plus other stuff
We had recently added code that intializes the buffer as the max buffer
size for the computer. The issue with this was when you took a new
measurement there would be old data left in the buffer from much
earlier that you would want to throw away. To make matters worse
in order to stop dropping packets at caltech I had to make the
default buffer for linux very big ~3000 packets. We were kind of
dealing with this by using the skip_packets keyword. but when the
buffer is large this becomes impractical because you dont want to
read in the 3000 new packets to get to the relavent data when doing
an IQ sweep. So I added function to clear the buffer of old packets
quickly and now it is called every time before taking data.

In addition I added keywords start_chan, end_chan and sub_folder
to saveDirfile_chanrangeIQ so that it can be called from some other
script without relying on real time user input. i.e so you can take
data while you are sleeping.

Also the indexing for packet_count was outdated in that function
so it was updated.

Finally the parallel process that writes the streaming to disk
is reniced so that it doesn't competed with the grabbing packet
process.