removed a couple conversion to MHz that were not correct
there were a few instance when sending the valon the frequency
command it was dividing by 10^6 when it shouldn't have been.
This was usually just after it finished a measurement and was
trying to reset the freuquency. So actually you wouldn't ever
notice it except that we where using a different frequency
source that shot back an error if the frequency range was
outside an aceptable range.