in saveSweeepData  moved saving of file outside while loop
in saveSweepData the saving of the data was happening after every
new packet and just overwriting the file each time. Since it was
always overwritten at the last packet you never noticed this in
the saved data. but I noticed it was making the iq sweeps hella
slow. so I moved the saving outside the while loop