Added in script for interactively tuneing power levels
I added in a file to that has a interactive plot for selecting
power levels for a tranfer fuction. It does its best to auto
predict the proper power levels, however this often fails
if you have a lot of collisions. So becuase of this you can
now sort throght the resonators at thier different power
levels and overide the automatic picked power level.